Add new functionalities and improve existing ones
This commit includes changes to several files and includes the addition of new functionalities, and the improvement of existing ones. Details include the introduction of enum Tipo, naming & typing changes to several variables, the renaming of the bh_provider to bh_controller, and some refactoring for better code structure and clarity. Furthermore, error handling has been added to the makeRows method. Specific changes to Widgets and UI elements have been made as well.