Updated dashboard functionalities and aesthetics
Refactored the code to improve the functionality of the dashboard notably changing the sorting logic, allowing the handlePageChange function to accommodate both forward and backward navigation, and tuning the response to the filters applied. Furthermore, aesthetics of the dashboard was enhanced by updating the application icons and modifying some interface elements.